[Guardian] Single hashi-guardian-init binary with subcommands
Replace the standalone provisioner-init binary with one
`hashi-guardian-init` binary exposing a `provisioner` subcommand; operator
init lands as a sibling subcommand. The two flows share the read + verify +
state-reconstruction core and differ only at the ends, so one entry point
over shared code fits better than separate binaries.
